\name{ImbTreeEntropy-package}
\alias{ImbTreeEntropy-package}
\docType{package}
\title{
  An R package for building entropy based classification trees on the imbalanced datasets
}
\description{
An R package for building binary and multiclass decision tree algorithms using generalized entropy functions, such as Renyi, Tsallis, Sharma-Mittal, Sharma-Taneja and Kapur, to measure impurity of a node. These are important extensions of the existing algorithms which usually employ Shannon entropy and the concept of information gain. Additionally, ImbTreeEntropy is able to handle imbalanced data which is a challenging issue in many practical applications. The package supports cost-sensitive learning by defining a misclassification cost matrix and weight sensitive learning. It accepts all types of attributes, including continuous, ordered and nominal. 
}

\author{
Krzysztof Gajowniczek

Maintainer: krzysztof_gajowniczek@sggw.edu.pl
}
\references{

}
\keyword{ package }
\seealso{
  \code{\link{ImbTreeEntropy}}, \code{\link{ImbTreeEntropyInter}}, \code{\link{PredictTree}}, \code{\link{PrintTree}}, \code{\link{PrintTreeInter}}, \code{\link{ExtractRules}}
}
\examples{
  \dontrun{
 
  }
}
