% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImbTreeEntropyInter.R
\name{PredictTree}
\alias{PredictTree}
\title{Predictions from an ImbTreeEntropy Object}
\usage{
PredictTree( tree, data )
}
\arguments{
\item{tree}{Fitted model object. This is assumed to be the result of some function that produces an object with the same named components as that returned by the ImbTreeEntropy or ImbTreeEntropyInter functions.}
\item{data}{Data frame of new examples to be predicted.}
}


\value{
Data frame with the probability of each class and predicted class label.
}
\seealso{
  \code{\link{ImbTreeEntropy}}, \code{\link{ImbTreeEntropyInter}}, \code{\link{PredictTree}}, \code{\link{PrintTree}}, \code{\link{PrintTreeInter}}, \code{\link{ExtractRules}}
}

\examples{
library("ImbTreeEntropy")
data(iris)
Tree <- ImbTreeEntropy(Y_name = "Species", 
                       X_names = colnames(iris)[-ncol(iris)], 
                       data = iris) 
Tree <- PredictTree(Tree, iris)
}
