% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImbTreeEntropyInter.R
\name{PrintTree}
\alias{PrintTree}
\title{Print an ImbTreeEntropy Object}
\usage{
PrintTree( tree )
}
\arguments{
\item{tree}{Fitted model object. This is assumed to be the result of some function that produces an object with the same named components as that returned by the ImbTreeEntropy function.}
}
\value{
Printed an ImbTreeEntropy based model.
}
\seealso{
  \code{\link{ImbTreeEntropy}}, \code{\link{ImbTreeEntropyInter}}, \code{\link{PredictTree}}, \code{\link{PrintTree}}, \code{\link{PrintTreeInter}}, \code{\link{ExtractRules}}
}

\examples{
library("ImbTreeEntropy")
data(iris)
Tree <- ImbTreeEntropy(Y_name = "Species", 
                       X_names = colnames(iris)[-ncol(iris)], 
                       data = iris) 
Tree <- PrintTree(Tree)
}
