% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImbTreeEntropyInter.R
\name{PrintTreeInter}
\alias{PrintTreeInter}
\title{Print an ImbTreeEntropyInter Object}
\usage{
PrintTreeInter( tree )
}
\arguments{
\item{tree}{Fitted model object. This is assumed to be the result of some function that produces an object with the same named components as that returned by the ImbTreeEntropyInter functions.}
}
\value{
Printed an ImbTreeEntropyInter based model.
}
\seealso{
  \code{\link{ImbTreeEntropy}}, \code{\link{ImbTreeEntropyInter}}, \code{\link{PredictTree}}, \code{\link{PrintTree}}, \code{\link{PrintTreeInter}}, \code{\link{ExtractRules}}
}

\examples{
library("ImbTreeEntropy")
data(iris)
Tree <- ImbTreeEntropyInter(Y_name = "Species", 
                            X_names = colnames(iris)[-ncol(iris)], 
                            data = iris) 
Tree <- PrintTreeInter(Tree)
}
